+++ schema/postgres/v1.10.0.sql
@@ -1,0 +1,1 @@
+ALTER TABLE "hit" ADD COLUMN "screen" varchar(20);